requirements.txt
requests==2.26.0
pandas==1.3.3
tqdm==4.62.3
sentence-transformers==2.2.0
tidb-vector==0.1.1
python-dotenv==0.19.1
SQLAlchemy==1.4.25
PyMySQL==1.0.2
torch==1.9.0
transformers==4.11.3
numpy==1.21.2